pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "locus-tag"
version = "0.4.0"
description = "A high-performance fiducial marker detector for robotics."
readme = "README.md"
authors = [{name = "Noé Fontana"}]
requires-python = ">=3.10"
license = {text = "MIT OR Apache-2.0"}
keywords = ["apriltag", "aruco", "fiducial", "marker", "detection", "robotics", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Rust",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Typing :: Typed",
]
dependencies = ["numpy>=2.0", "pydantic>=2.0"]

[project.urls]
Homepage = "https://github.com/NoeFontana/locus-tag"
Source = "https://github.com/NoeFontana/locus-tag"
Documentation = "https://noefontana.github.io/locus-tag"
Issues = "https://github.com/NoeFontana/locus-tag/issues"
Changelog = "https://github.com/NoeFontana/locus-tag/blob/main/CHANGELOG.md"

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "locus"
python-source = "crates/locus-py"
manifest-path = "crates/locus-py/Cargo.toml"
include = ["crates/locus-core/data/**/*", "LICENSE-APACHE", "LICENSE-MIT"]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "UP", "SIM", "I"] # Pyflakes, pycodestyle, bugbear, pyupgrade, simplify, isort
ignore = ["E501", "B006", "B008"] # Line length handled by formatter

[tool.pyright]
pythonVersion = "3.10"
include = ["crates/locus-py", "tools", "tests", "scripts", "examples"]
exclude = ["**/.venv", "**/target", "**/__pycache__", "**/.mypy_cache", "**/site"]
extraPaths = ["crates/locus-py"]
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
# Unstubbed third-party deps (datasets, rerun, cv2, pupil_apriltags, pyarrow):
# pyright has no per-module-glob equivalent of mypy's [[overrides]], so we
# silence missing-stub noise globally while keeping import-resolution errors on.
reportMissingTypeStubs = "none"
reportMissingImports = "error"
# Selective strictness ramp — high-value reports beyond `standard` defaults.
reportUnnecessaryTypeIgnoreComment = "warning"
reportUnusedImport = "warning"

[dependency-groups]
dev = [
    "maturin>=1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "typer>=0.12.0",
    "jsonschema>=4.0.0",
]
lint = [
    "ruff>=0.14.14",
]
types = [
    "basedpyright>=1.20",
    "pandas-stubs>=2.0.0",
    "types-tqdm>=4.0.0",
    "types-jsonschema",
    "types-requests",
    "types-PyYAML",
    "types-setuptools",
]
bench = [
    "opencv-python-headless>=4.0.0",
    "rerun-sdk>=0.9.0",
    "matplotlib>=3.0.0",
    "datasets[vision]",
    "pupil-apriltags",
    "typer>=0.12.0",
    "pyarrow>=15.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
etl = [
    "huggingface-hub>=0.19.0",
    "tqdm>=4.66.0",
]
